Refactor to seperate PListMLE and ListMLE
